wafel_window
Trait AppEnv

pub trait AppEnv {
    // Required methods
    fn config(&self) -> &AppConfig;
    fn first_run(&self) -> bool;
    fn fps(&self) -> f32;
    fn mspf(&self) -> f32;
    fn input(&self) -> &Input;
    fn egui_ctx(&self) -> &Context;
    fn draw_viz(&self, scene: VizScene);
    fn take_recent_panic_details(&self) -> Option<String>;
}

Trait defining the interaction between a windowed application and the window.

Required Methods

fn config(&self) -> &AppConfig

The config that was used when running the application.

fn first_run(&self) -> bool

True if this is the first time that the app callback has been called.

fn fps(&self) -> f32

A recent fps measurement.

fn mspf(&self) -> f32

A recent mspf measurement.

fn input(&self) -> &Input

The keyboard/mouse input state.

fn egui_ctx(&self) -> &Context

The egui context.

fn draw_viz(&self, scene: VizScene)

Adds a [wafel_viz] visualization to the window.

fn take_recent_panic_details(&self) -> Option<String>

Return details of the most recent panic caught by the panic handler.

This method also clears the panic details.