Trait wafel_window::WindowEnv

pub trait WindowEnv {
    // Required methods
    fn config(&self) -> &Config;
    fn fps(&self) -> f32;
    fn mspf(&self) -> f32;
    fn egui_ctx(&self) -> &Context;
    fn draw_viz(&self, render_data: VizRenderData);
    fn take_recent_panic_details(&self) -> Option<String>;
}

Trait defining the interaction between a windowed application and the window.

Required Methods

fn config(&self) -> &Config

The config that was used when running the application.

fn fps(&self) -> f32

A recent fps measurement.

fn mspf(&self) -> f32

A recent mspf measurement.

fn egui_ctx(&self) -> &Context

The egui context.

fn draw_viz(&self, render_data: VizRenderData)

Adds a [wafel_viz] visualization to the window.

This method is only available when the wafel_viz feature is enabled.

fn take_recent_panic_details(&self) -> Option<String>

Return details of the most recent panic caught by the panic handler.

This method also clears the panic details.