Crate wafel_window

A crate for creating a window and rendering with wgpu, with support for egui and wafel_viz.

This crate also initializes logging to a file and stderr.

Example

// This prevents the console window from appearing on Windows in release mode.
#![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]

let config = wafel_window::AppConfig::new().with_title("Minimal example");

wafel_window::run(&config, move |env| {
    let ctx = env.egui_ctx();

    egui::CentralPanel::default().show(ctx, |ui| {
        ui.label(format!("{:#?}", env.config()));
        ui.label(format!("{:.3} mspf = {:.1} fps", env.mspf(), env.fps()));
    });
});

Features

This crate has one optional feature, which is disabled by default:

image: Enables loading icons from .ico files (Windows only).

Structs

AppConfig

Configuration for the window and application environment.

DragState

Helper struct which tracks the state of a mouse drag.

Input

Access to keyboard state and events.

Enums

MouseButton

Describes a button of a mouse controller.

VirtualKeyCode

Symbolic name for a keyboard key.

Traits

AppEnv

Trait defining the interaction between a windowed application and the window.

Functions

run

Initializes logging, opens a window and runs the application.